+++ junit5-extensions/junit5-extensions-no-andhow-dependency/src/test/java/org/yarnandtail/andhow/junit5/ext/ExtensionBaseTest.java
@@ -1,0 +1,29 @@
+package org.yarnandtail.andhow.junit5.ext;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ExtensionBaseTest {
+
+	@BeforeEach
+	void setUp() {
+	}
+
+	@Test
+	void getPerTestClassStore() {
+	}
+
+	@Test
+	void getPerTestNamespace() {
+	}
+
+	@Test
+	void getPerTestMethodStore() {
+	}
+
+	@Test
+	void getPerTestMethodNamespace() {
+	}
+}
