+++ junit5-extensions/junit5-extensions-with-andhow-dependency/src/test/java/org/yarnandtail/andhow/junit5/ext/ConfigFromFileExtUsageTest.java
@@ -1,0 +1,14 @@
+package org.yarnandtail.andhow.junit5.ext;
+
+import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.mockito.*;
+import org.yarnandtail.andhow.*;
+import org.yarnandtail.andhow.testutil.AndHowTestUtils;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.times;
+
+class ConfigFromFileExtUsageTest {
+
+}
